cleaning up dotfiles

+++ teardown.sh
@@ -36,13 +36,22 @@ fi
 
 # Stop and remove the Nix daemon services
 sudo launchctl unload /Library/LaunchDaemons/org.nixos.nix-daemon.plist
-sudo rm /Library/LaunchDaemons/org.nixos.nix-daemon.plist
+
+if [ -f "/Library/LaunchDaemons/org.nixos.nix-daemon.plist" ]; then
+  sudo rm /Library/LaunchDaemons/org.nixos.nix-daemon.plist
+fi
+
 sudo launchctl unload /Library/LaunchDaemons/org.nixos.darwin-store.plist
-sudo rm /Library/LaunchDaemons/org.nixos.darwin-store.plist
+
+if [ -f "/Library/LaunchDaemons/org.nixos.darwin-store.plist" ]; then
+  sudo rm /Library/LaunchDaemons/org.nixos.darwin-store.plist
+fi
 
 # Remove the nixbld group and the _nixbuildN users
-sudo dscl . -delete /Groups/nixbld
-for u in $(sudo dscl . -list /Users | grep _nixbld); do sudo dscl . -delete /Users/$u; done
+if [ sudo /usr/bin/dscl . -list /Groups | grep nixbld ]; then
+  sudo dscl . -delete /Groups/nixbld
+  for u in $(sudo dscl . -list /Users | grep _nixbld); do sudo dscl . -delete /Users/$u; done
+fi
 
 # Edit fstab using sudo vifs to remove the line mounting the Nix Store volume on /nix, which looks like
 # UUID=<uuid> /nix apfs rw,noauto,nobrowse,suid,owners
@@ -63,6 +72,6 @@ fi
 sudo rm -rf /etc/nix /var/root/.nix-profile /var/root/.nix-defexpr /var/root/.nix-channels ~/.nix-profile ~/.nix-defexpr ~/.nix-channels ~/.local/share/nix ~/.local/state/nix ~/.cache/nix
 
 # Remove the Nix Store volume
-sudo diskutil apfs deleteVolume /nix
+sudo /usr/sbin/diskutil apfs deleteVolume /nix
 
-diskutil list
+#diskutil list
